functions/requirements.txt
firebase-functions
firebase-admin
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
Flask-Cors
